Communication between boards
Master has changes in i2c_master.c's readreg and writereg to
demonstrate how to communicate with the slave board
Slave2 has changes in lufa.c with general functions that can be moved
to i2c_master.c.  This does not have error correction as I am having
a lot of confusion between what the other states mean and when to
send and ack/nack back.